Allow non-optional options-object-like parameters
Previously, a parameter name with a period (as in an options object
property) was only recognized if the portion preceding the period
(i.e., the variable name of the property's parent object) also existed
as a parameter. This change recognizes object property parameters
without the corresponding parent object also existing as a parameter.